Home task №4 (draft)

+++ MetricsManager/MetricsAgent/Models/NetworkMetric.cs
@@ -6,6 +6,6 @@ namespace MetricsAgent.Models
     {
         public int Id { get; set; }
         public int Value { get; set; }
-        public double Time { get; set; }
+        public long Time { get; set; }
     }
 }
